chore(slack_app): keep every eval scorer in one module
The product had two scorer modules split by suite rather than by concern, so a
new suite had to pick one or add a third. Fold them together: the suites grade a
handful of recurring shapes, and reading the model-override pair next to the
answer judges is what makes the asymmetry between them legible.

Renames EXPECTATION_KEY to MODEL_OVERRIDE_KEY — sitting beside the existing
EXPECTED_KEY, the old name read as a near-duplicate of it.
